+++ mobile/android/android-components/components/lib/fetch-okhttp/src/test/java/mozilla/components/lib/fetch/okhttp/OkHttpFetchTestCases.kt
@@ -7,9 +7,13 @@ package mozilla.components.lib.fetch.okhttp
 import mozilla.components.concept.fetch.Client
 import org.junit.Assert.assertTrue
 import org.junit.Test
+import org.junit.runner.RunWith
+import org.robolectric.RobolectricTestRunner
+import org.robolectric.RuntimeEnvironment
 
+@RunWith(RobolectricTestRunner::class)
 class OkHttpFetchTestCases : mozilla.components.tooling.fetch.tests.FetchTestCases() {
-    override fun createNewClient(): Client = OkHttpClient(okhttp3.OkHttpClient())
+    override fun createNewClient(): Client = OkHttpClient(okhttp3.OkHttpClient(), RuntimeEnvironment.application)
 
     // Inherits test methods from generic test suite base class